refactor: 계산 메서드를 CustomTextCalculator 클래스로 분리
